+++ tests/Unit/JsonMapperTest.php
@@ -8,52 +8,23 @@ use JsonMapper\Handler\PropertyMapper;
 use JsonMapper\JsonMapper;
 use JsonMapper\JsonMapperInterface;
 use JsonMapper\Middleware\AbstractMiddleware;
+use JsonMapper\Tests\Implementation\IsCalledHandler;
+use JsonMapper\Tests\Implementation\IsCalledMiddleware;
 use JsonMapper\ValueObjects\PropertyMap;
 use JsonMapper\Wrapper\ObjectWrapper;
 use PHPUnit\Framework\TestCase;
 
 class JsonMapperTest extends TestCase
 {
-    /** @var callable */
+    /** @var IsCalledHandler */
     private $handler;
-    /** @var AbstractMiddleware */
+    /** @var IsCalledMiddleware */
     private $middleware;
 
     protected function setUp(): void
     {
-        $this->handler = new class {
-            /** @var bool */
-            private $called = false;
-
-            public function __invoke(): void
-            {
-                $this->called = true;
-            }
-
-            public function isCalled(): bool
-            {
-                return $this->called;
-            }
-        };
-
-        $this->middleware = new class extends AbstractMiddleware{
-            /** @var bool */
-            private $called = false;
-
-            public function isCalled(): bool
-            {
-                return $this->called;
-            }
-
-            public function handle(
-                \stdClass $json,
-                ObjectWrapper $object,
-                PropertyMap $propertyMap,
-                JsonMapperInterface $mapper
-            ): void {
-                $this->called = true;
-            }
-        };
+        $this->handler = new IsCalledHandler();
+        $this->middleware = new IsCalledMiddleware();
     }
 
     /**